Add --enable-graphite option to build script
GCC 4.5 and up supports graphite optimization, and cloog and ppl
are required.  This change attempts to support some combinations
of ppl and cloog/cloog-ppl libraries.  Since cloog and ppl are
written in C++, it implies we might suffer from C++ ABI breakage
problems when host environment changes, that is a well-known issue
in GNU/Linux distributions.  Thus, we have to link to static C++
runtime library (libstdc++) to avoid depending on the host version.

Default: no

Change-Id: I0a69712379a4e2dc30034a2226a15717a9c121b7