Define correct default gcc target march/mtune/mfpmath flags for x86
This is required to enforce correct x86 -march/-mtune flags which are
currently hardcoded in gcc (see gcc/config/i386/gnu-user.h). The problem
is that gcc hardcode doesn't work, it is overridden by generic values.
It can be seen with -### option for any compilation.
This patch adds correct configure flags and after it is merged
-march/-mtune flags in gcc can be removed.

Default -mfpmath can't be seen in -### but can be configured as well to
eliminate hardcode.

Also cleanup old flags which didn't work anyway. target_cpu is i686 and
not x86.

Change-Id: I43ea7034e855629933da2d1bede90cd74a238861
Signed-off-by: Pavel Chupin <pavel.v.chupin@intel.com>